requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langchain-openai
langchain-text-splitters
fastembed
faiss-cpu
python-dotenv
pypdf
